+++ flutter_sound/CHANGELOG.md
@@ -1,4 +1,4 @@
 ## 0.1.20
 
-- Please [look to this](/CHANGELOG.md)
+- Please [look to this](../CHANGELOG.md)